Remove Secret Key from comitted settings file
Wheeeeeee!  Turns out Django automatically generates a secret key
and stores it in settings.py.  This is insecure.  Instead, we
should load the secret key from an env var.  Sadly, the book does
not make note of this until the point where you automate the
deployment, so this got committed very early on.